requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
Pillow==5.0.0
graphene_django>=2.0
django-cors-headers==3.8.0
django-model-utils>=4.1.1